+++ sentry-android-core/src/main/java/io/sentry/android/core/AppStartState.java
@@ -25,7 +25,7 @@ public final class AppStartState {
 
   private AppStartState() {}
 
-  static @NotNull AppStartState getInstance() {
+  public static @NotNull AppStartState getInstance() {
     return instance;
   }
 
@@ -51,7 +51,7 @@ public final class AppStartState {
     return appStartEndMillis - appStartMillis;
   }
 
-  boolean isColdStart() {
+  public boolean isColdStart() {
     return coldStart;
   }
 
@@ -60,7 +60,7 @@ public final class AppStartState {
   }
 
   @Nullable
-  Date getAppStartTime() {
+  public Date getAppStartTime() {
     return appStartTime;
   }