+++ sentry/src/main/java/io/sentry/Hub.java
@@ -620,7 +620,8 @@ public final class Hub implements IHub {
       final @NotNull TransactionContext transactionContext,
       final @Nullable CustomSamplingContext customSamplingContext,
       final boolean bindToScope) {
-    return createTransaction(transactionContext, customSamplingContext, bindToScope, null, false);
+    return createTransaction(
+        transactionContext, customSamplingContext, bindToScope, null, false, null);
   }
 
   @ApiStatus.Internal
@@ -631,7 +632,7 @@ public final class Hub implements IHub {
       boolean bindToScope,
       @Nullable Date startTimestamp) {
     return createTransaction(
-        transactionContext, customSamplingContext, bindToScope, startTimestamp, false);
+        transactionContext, customSamplingContext, bindToScope, startTimestamp, false, null);
   }
 
   @ApiStatus.Internal
@@ -641,9 +642,15 @@ public final class Hub implements IHub {
       final @Nullable CustomSamplingContext customSamplingContext,
       final boolean bindToScope,
       final @Nullable Date startTimestamp,
-      final boolean waitForChildren) {
+      final boolean waitForChildren,
+      final @Nullable TransactionFinishedCallback transactionFinishedCallback) {
     return createTransaction(
-        transactionContexts, customSamplingContext, bindToScope, startTimestamp, waitForChildren);
+        transactionContexts,
+        customSamplingContext,
+        bindToScope,
+        startTimestamp,
+        waitForChildren,
+        transactionFinishedCallback);
   }
 
   private @NotNull ITransaction createTransaction(
@@ -651,7 +658,8 @@ public final class Hub implements IHub {
       final @Nullable CustomSamplingContext customSamplingContext,
       final boolean bindToScope,
       final @Nullable Date startTimestamp,
-      final boolean waitForChildren) {
+      final boolean waitForChildren,
+      final @Nullable TransactionFinishedCallback transactionFinishedCallback) {
     Objects.requireNonNull(transactionContext, "transactionContext is required");
 
     ITransaction transaction;
@@ -674,7 +682,13 @@ public final class Hub implements IHub {
       boolean samplingDecision = tracesSampler.sample(samplingContext);
       transactionContext.setSampled(samplingDecision);
 
-      transaction = new SentryTracer(transactionContext, this, startTimestamp, waitForChildren);
+      transaction =
+          new SentryTracer(
+              transactionContext,
+              this,
+              startTimestamp,
+              waitForChildren,
+              transactionFinishedCallback);
     }
     if (bindToScope) {
       configureScope(scope -> scope.setTransaction(transaction));