+++ sentry/src/main/java/io/sentry/SpanFinishedCallback.java
@@ -2,11 +2,11 @@ package io.sentry;
 
 import org.jetbrains.annotations.NotNull;
 
-interface SpanListener {
+interface SpanFinishedCallback {
   /**
    * Called when observed span finishes.
    *
    * @param span the span that has finished.
    */
-  void onSpanFinished(final @NotNull Span span);
+  void execute(final @NotNull Span span);
 }