+++ sentry/src/main/java/io/sentry/protocol/MeasurementValue.java
@@ -1,6 +1,7 @@
 package io.sentry.protocol;
 
 import org.jetbrains.annotations.ApiStatus;
+import org.jetbrains.annotations.TestOnly;
 
 @ApiStatus.Internal
 public final class MeasurementValue {
@@ -9,5 +10,10 @@ public final class MeasurementValue {
 
   public MeasurementValue(final float value) {
     this.value = value;
+  }
+
+  @TestOnly
+  public float getValue() {
+    return value;
   }
 }